fix: percentile() aggregate emits Float instead of Decimal for percentile argument (#5301)
`buildAggregateExpression` in `QueryBuilderAggregateOperator_Percentile` was
constructing the percentile `PrimitiveInstanceValue` with `PrimitiveType.NUMBER`
(the abstract numeric supertype). The V1 value-specification transformer has no
direct wire representation for `Number` and intentionally falls back to `Decimal`,
producing `{"_type":"decimal","value":0.8}` → PURE grammar `0.8D`. The Engine
correctly rejects this because the registered function signature is:

```
percentile(Number[*], Float[1], Boolean[1], Boolean[1]): Number[0..1]
```

`Decimal` and `Float` are distinct concrete subtypes of `Number` and are not
interchangeable in the PURE type system.

**Fix:** change the generic type from `PrimitiveType.NUMBER` to `PrimitiveType.FLOAT`
so the transformer takes the `FLOAT` branch → `V1_CFloat` →
`{"_type":"float","value":0.8}` → PURE grammar `0.8` → matches `Float[1]`.

### Files changed

| File | Change |
|---|---|
| `…/QueryBuilderAggregateOperator_Percentile.ts` | `PrimitiveType.NUMBER` → `PrimitiveType.FLOAT` (one line) |
| `…/TEST_DATA__QueryBuilder_Generic.ts` | Updated test data: `"_type":"decimal"` → `"_type":"float"` |
| `…/QueryBuilderFetchStructure.test.tsx` | Added serialization-check test + load-back roundtrip test |
| `…/QueryBuilderProcessingRoundtrip.test.ts` | Added `percentile()` roundtrip case |
| `.changeset/percentile-float-fix.md` | `patch` changeset for `@finos/legend-query-builder` |

+++ packages/legend-query-builder/src/stores/__tests__/TEST_DATA__QueryBuilder_Generic.ts
@@ -2023,7 +2023,7 @@ export const TEST_DATA__projectionWithPercentileAggregation = {
                           name: 'x',
                         },
                         {
-                          _type: 'decimal',
+                          _type: 'float',
                           value: 0.45,
                         },
                         {
@@ -2083,7 +2083,7 @@ export const TEST_DATA__projectionWithPercentileAggregation = {
                           name: 'x',
                         },
                         {
-                          _type: 'decimal',
+                          _type: 'float',
                           value: 0.36,
                         },
                       ],
